Refactor Kafka client API
