+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kafka-bridge"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Stephen Blum <blum.stephen@gmail.com>"]
 edition = "2018"
 readme = "readme.md"
